requirements.txt
click>=8.0.0
pyyaml>=6.0
requests>=2.28.0
gitpython>=3.1.30
pygithub>=1.58.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
openai>=1.0.0
beautifulsoup4>=4.12.0
